pyproject.toml
[project]
name = "langgraph-architecture"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.11.0.post1",
    "ipykernel>=6.30.1",
    "ipython>=9.4.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.3",
    "langsmith>=0.4.11",
    "openai>=1.98.0",
    "pypdf>=5.9.0",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "sentence-transformers>=5.0.0",
    "streamlit>=1.47.1",
]
